Android: Allow the track to be replayed (#185)
When the track completes, abandon the audio focus but don't release the
media player. This allows the track to be played again by calling
seekTo(0) and play(), like you can when running the plugin with iOS.

Previously trying to call methods such as seekTo or play after the track
completed resulted in {"player":{},"error":-38,"extra":0}. The only way
to play the track a second time was to call the init method again.